requirements.txt
asgiref==3.7.2
Django==4.2.4
Django-Verify-Email==2.0.3
Pillow==10.0.0
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
